refactored, accidentally testing multiple server files
